Open file and load data

+++ knn/knn_create_model.py
@@ -1,4 +1,5 @@
 import csv, json, nltk, string, sys
+from os import path
 
 
 def error(name):
@@ -20,37 +21,74 @@ def tokenize(value):
         # Remove punctuations and make all words lower case
         words.append(word)
     return words
+   
+# Take train corpus --> output to CSV
+
+class KnnModel:
+    def __init__(self,corpus, outputFile):
+        self.corpus = corpus
+        self.document = []
+        self.IdfLines = []
+        self.VectorLines = []
+        self.outputFile = outputFile
+    
+    def vectorize(self):
+        for document in self.corpus:
+            #represent as a vector using ltn wieght scheme
+            self.document.append(document)
+
+    def getIdf(self):
+        for _ in self.document:
+            self.outputFile.writerow(["IDF", "term", "IDFWeight"])
+
+    def getVector(self):
+        #for class in model:
+        self.outputFile.writerow(["Vectpr", "class", "STRING REPRESENTATION OF VECTOR FOR CLASS"])
 
 
 def main():
     # Get the arguments and validate the number of arguments
     arguments = sys.argv
     if len(arguments) != 3:
-        error("Invalid arguments")
+        error("Invalid number of arguments")
 
     inputName = arguments[1]
     outputName = arguments[2]
-
-    print(inputName, outputName)
 
     # Open the input json file for read
     try:
         inputFile = open(inputName, 'r')
     except IOError:
-        error('Invalid file arguments')
+        error('Invalid input file argument')
+
+    # Open the output file for write
+    if path.exists(outputName):
+        while True:
+            user = input('Do you want to replace the file: ' + outputName + '?\n').lower()
+            if user in ['no', 'n']:
+                print('Mission aborted')
+                inputFile.close()
+                return
+            elif user in ['yes', 'y']:
+                break
+
+    try:
+        outputFile = open(outputName, 'w', newline='')
+    except IOError:
+        error('Invalid putput file argument')
 
     # Load and parse json data
     inputData = json.load(inputFile)
     inputFile.close()
-    print(inputData)
 
-    # Open the output file for write
-    try:
-        outputFile = open(outputName, 'w', newline='')
-    except IOError:
-        error('Invalid file arguments')
     theWriter = csv.writer(outputFile, delimiter='\t')
     theWriter.writerow(['ID', 'normalized weight'])
+
+    knn = KnnModel(inputData, theWriter)
+    knn.vectorize()
+    knn.getIdf()
+    knn.getVector()
+
     outputFile.close()
 
 
@@ -58,23 +96,3 @@ if __name__ == "__main__":
     main()
 
     print('\nDone\n')
-   
-#take train corpus --> output to CSV
-
-class KnnModel:
-    def __init__(self,corpus):
-        self.corpus = corpus
-        self.IdfLines = []
-        self.VectorLines = []
-    
-    def vectorize(self):
-        for document in self.corpus:
-            #represent as a vector using ltn wieght scheme
-         
-    def getIdf(self):
-        for trerm in document:
-            file.write("IDF", Term,, IDFWeight)
-
-    def getVector(self):
-        #for class in model:
-        file.write("Vectpr", class, "STRING REPRESENTATION OF VECTOR FOR CLASS')")
